more RP

+++ src/main/java/org/ranch/ballshack/gui/legacy/LegacyGuiUtil.java
@@ -2,7 +2,7 @@ package org.ranch.ballshack.gui.legacy;
 
 public class LegacyGuiUtil {
 	public static boolean mouseOverlap(int mouseX, int mouseY, int x, int y, int width, int height) {
-		return mouseOverlap( (double) mouseX, (double) mouseY, x, y, width, height);
+		return mouseOverlap(mouseX, (double) mouseY, x, y, width, height);
 	}
 
 	public static boolean mouseOverlap(double mouseX, double mouseY, int x, int y, int width, int height) {
